Split server api handler
