+++ src/common/ButtonWithConfirm.tsx
@@ -3,9 +3,8 @@ import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogTitle from '@mui/material/DialogTitle';
-import CircularProgress from '@mui/material/CircularProgress';
-import Box from '@mui/material/Box';
 import { Fragment, memo, useCallback, useState } from 'react';
+import { Loading } from './LoadingAnimation';
 
 interface IButtonWithConfirm {
   buttonLabel: string;
@@ -15,11 +14,7 @@ interface IButtonWithConfirm {
   okLabel?: string;
   cancelLabel?: string;
 }
-const Loading = () => (
-  <Box sx={{ display: 'flex', justifyContent: 'center' }}>
-    <CircularProgress />
-  </Box>
-);
+
 function _ButtonWithConfirm(props: IButtonWithConfirm) {
   const [open, setOpen] = useState(false);
   const [loading, setLoading] = useState(false);
