+++ tljh_repo2docker/app.py
@@ -16,6 +16,7 @@ from .builder import BuildHandler
 from .environments import EnvironmentsHandler
 from .logs import LogsHandler
 from .servers import ServersHandler
+from .servers_api import ServersAPIHandler
 
 if os.environ.get("JUPYTERHUB_API_TOKEN"):
     from jupyterhub.services.auth import HubOAuthCallbackHandler
@@ -157,7 +158,6 @@ class TljhRepo2Docker(Application):
         handlers = []
         static_path = str(HERE / "static")
         server_url = url_path_join(self.service_prefix, r"servers")
-        environments_url = url_path_join(self.service_prefix, r"environments")
         handlers.extend(
             [
                 (
@@ -177,7 +177,11 @@ class TljhRepo2Docker(Application):
                 (self.service_prefix, web.RedirectHandler, {"url": server_url}),
                 (server_url, ServersHandler),
                 (
-                    environments_url,
+                    url_path_join(self.service_prefix, r"api/servers"),
+                    ServersAPIHandler,
+                ),
+                (
+                    url_path_join(self.service_prefix, r"environments"),
                     EnvironmentsHandler,
                 ),
                 (url_path_join(self.service_prefix, r"api/environments"), BuildHandler),
